+++ lib/flavor_config.dart
@@ -2,4 +2,5 @@ library flavor_config;
 
 export 'package:flavor_config/src/config/flavor_config.dart';
 export 'package:flavor_config/src/banner/flavor_banner.dart';
+export 'package:flavor_config/src/dialog/device_info_dialog.dart';
 export 'package:flavor_config/src/utils/device_utils.dart';